requirements.txt
torch
torchvision
tqdm
Pillow
opencv-python

git+https://github.com/FacePerceiver/facer.git@main
